user register functionality

+++ base/views.py
@@ -1,17 +1,25 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
-from .models import Room, Topic
+from django.contrib.auth.decorators import login_required
+from .models import Room, Topic, Message
 from django.contrib.auth.models import User
 from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models import Q
 
 # Create your views here.
 
 def login_user(request):
 
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -27,12 +35,35 @@ def login_user(request):
         else:
             messages.error(request, 'Username or Password Does not Exist')
 
-    context = {}
+    context = {
+        "page": page,
+    }
     return render(request, 'base/login_register.html', context)
 
 def logout_user(request):
     logout(request)
     return redirect('home')
+
+def register_user(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occured during registration")
+
+    context = {
+        "page": page,
+        "form": form,
+    }
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -54,12 +85,15 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all().order_by('-created')
 
     context = {
-        'room': room
+        'room': room,
+        'room_messages': room_messages
     }
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
 
@@ -74,9 +108,13 @@ def create_room(request):
     }
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -89,8 +127,12 @@ def update_room(request, pk):
     }
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here!")
 
     if request.method == 'POST':
         room.delete()
